% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_rule.R
\name{remove_rule}
\alias{remove_rule}
\title{Removes an existing rule}
\usage{
remove_rule(name = NULL, all = FALSE)
}
\arguments{
\item{name}{character. Specified name/s of the rule to be removed.}

\item{all}{logical. Should all the rules be removed?}
}
\value{
Return a message confirming the removal.
}
\description{
\code{remove_rule} is a function to removes the existing rules. Can remove several rules at once.
}
\details{
If all = TRUE is used, the name parametr name will not be evaluated at all.
}
\examples{
\dontrun{
remove_rule(name = "test1")
# removes rule called "test1"

remove_rule(all = TRUE)
# removes all rules
}
}
\author{
Michal Kubista
}
