% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rule.R
\name{show_rule}
\alias{show_rule}
\title{Shows existing rule}
\usage{
show_rule(name, all = TRUE)
}
\arguments{
\item{name}{character. Specified name of the rule to be shown.}

\item{all}{logical. Should all the rules be shown?}
}
\value{
Returns a data.frame of defined rules.
}
\description{
\code{show_rule} is a function to show defined rules. Can show all or several rules at once.
}
\details{
To see all the rules, it is enough to use \code{show_rule}, alternatively you can specify the names of the rules to be shown. If all = TRUE (default value) is used, the name parametr name will not be evaluated at all.
}
\examples{
\dontrun{
show_rule(name = "test1")
# show rule called "test1"

show_rule()
# show all rules, equivalent to show_rule(all= TRUE)
}
}
\author{
Michal Kubista
}
