% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rule.R
\name{update_rule}
\alias{update_rule}
\title{Updates an existing rule}
\usage{
update_rule(name, update)
}
\arguments{
\item{name}{character. Specified name of the rule.}

\item{update}{character. Named vector with parameters as names containing updated parameters.}
}
\value{
Return a message confirming the update of specified rule.
}
\description{
\code{update_rule} is a function to update the existing rules. Can update several arguments for a single rule.
}
\details{
For more details on format of rules please go to \code{\link{write_rule}}.
}
\examples{
\dontrun{
update_rule(name = "test1", update = c(x = "mtcars$cyl", type = "def"))
}
}
\author{
Michal Kubista
}
