% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify}
\alias{verify}
\title{Verifies specfified rules}
\usage{
verify(name = NULL, all = FALSE)
}
\arguments{
\item{name}{character. Specified name/s of the rule to be removed.}

\item{all}{logical. Should all the rules be removed?}
}
\value{
Returns a list of two objects:
\itemize{
\item{rules} table of all verified rules and results of the verification
\item{unsolved} list of vectors/tables, containing problematic values, or their row_indexes (rulename_index)
}
}
\description{
\code{verify} is a function used to verify your defined rules. Can verify several rules at once.
}
\details{
If all = TRUE is used, the name parametr name will not be evaluated at all.
}
\examples{
\dontrun{
verify(name = c("a", "b"))
verify(all = TRUE)
}
}
\author{
Michal Kubista
}
