% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_class_and_typeof.R
\name{verify_class_and_typeof}
\alias{verify_class_and_typeof}
\title{Verify class and type of a test value.}
\usage{
verify_class_and_typeof(
  tested_value,
  required_length,
  tested_value_name,
  required_classes,
  required_typeof
)
}
\arguments{
\item{tested_value}{the value of what is being tested, e.g. the variable loaded in as a function argument.}

\item{required_length}{Required length of the tested_value}

\item{tested_value_name}{the name of the variable, input as a character, e.g. if the value tested was first_derivative, then it's tested_value_name would be "first_derivative".}

\item{required_classes}{character vector of required classes for tested_value to possess.}

\item{required_typeof}{character vector of required types for tested_value to possess.}
}
\value{
The return value is the identity of tested_value if all assertions pass; otherwise, an assertion error in assert_that will trigger.
}
\description{
This function is a much more compact way of doing many common type-checking assertions for a
a function argument.
}
