% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rule.R
\name{write_rule}
\alias{write_rule}
\title{Write new rule to be verified.}
\usage{
write_rule(name, x, type, y, def , result, other)
}
\arguments{
\item{name}{character. Name of the rule. Needs to be unique for every rule.}

\item{x}{object. Data against which the rule is to be verified.}

\item{type}{character. Possible types are:
\itemize{
 \item{\code{integ} checks for presence of all unique values of x in y}
 \item{\code{summary} checks summarised data against original, more detailed data}
 \item{\code{na} checks the missing values}
 \item{\code{numeric} checks outlying (user defined) numeric values}
 \item{\code{def} any other type, needs to be defined in def parametr}.
}}

\item{y}{vector. Used in \code{type} \code{integrity} (y is the reference table) or
\code{summary} (y is the summarised table).}

\item{def}{character. Used for user defined rules using any function in evnironment, accepts only he name of the function, e.g. "sum".}

\item{result}{numeric. Used for \code{type} \code{numeric} or \code{def} to specify bounds or expected value.}

\item{other}{character. Used to specify other arguments passed into the user defined function, e.g. "na.rm = T".}
}
\value{
Returns a message confirming the creation of new rule.
}
\description{
\code{write_rule} is a function to help you write new rules. Can write several rules at once using vectors.
}
\details{
You can create several rules at once assigning vectors into the parametrs. When the \code{length} of any parameter is longer than one the function uses cbind to create a data.frame (will repeat values of vectors with smaller size). Names of the rules need to be unique.
}
\examples{
\dontrun{
write_rule(name = "test1", x = mtcars$mpg, type = "def", def = c("mean", "median"),
           result = 19.2, other ="na.rm = T")
# verify, that the mean and median of mtcars$mpg is 19.2, omitting the missing values

write_rule(name = "integrity1", x = data$cities, type = "integ", y = "ref$cities")
# verify, that all of the cities used in my data are in the reference table

}
}
\author{
Michal Kubista
}
